Move backtransform to its own function
